Exercise 4.21

+++ backend/controllers/blogController.js
@@ -19,8 +19,7 @@ const getAllBlogs = app.get('/api/blogs', async (request, response) => {
   try {
     const blogs = await Blog.find({});
     response.json(blogs);
-  }
-  catch (error) {
+  } catch (error) {
     console.log(error);
   }
 })
@@ -45,13 +44,13 @@ const createBlog = app.post('/api/blogs', async (request, response) => {
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' });
+    return response.status(401).json({error: 'token missing or invalid'});
   }
 
   const user = await User.findById(decodedToken.id);
 
   if (!body.title || !body.url) {
-    return response.status(400).json({ error: 'title or url missing' });
+    return response.status(400).json({error: 'title or url missing'});
   }
 
   try {
@@ -74,20 +73,34 @@ const createBlog = app.post('/api/blogs', async (request, response) => {
     await user.save();
 
     response.status(201).json(savedBlog);
-  }
-  catch (error) {
+  } catch (error) {
     console.log(error);
-    response.status(500).json({ error: "An Error occurred while creating the blog" });
+    response.status(500).json({error: "An Error occurred while creating the blog"});
   }
 })
 
 // Delete a blog
 const deleteBlog = app.delete('/api/blogs/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'});
+  }
+
+  const user = await User.findById(decodedToken.id);
+  if (!user) {
+    return response.status(401).json({error: 'user not found'});
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(404).send({error: 'blog not found'});
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({error: 'unauthorized user'});
+  }
+
   try {
-    const blog = await Blog.findById(request.params.id);
-    if (!blog) {
-      return response.status(404).send({ error: 'blog not found' });
-    }
     await Blog.findByIdAndDelete(request.params.id);
     response.status(204).end();
   } catch (error) {
@@ -107,7 +120,7 @@ const updateBlog = app.put('/api/blogs/:id', async (request, response) => {
   }
 
   try {
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true});
     response.json(updatedBlog);
   } catch (error) {
     console.log(error);
